requirements.txt
#requirements.txt

python-dotenv
google-cloud-vision
